chore: domain, view, controller 파일 위치 변경
