+++ app/src/main/java/woowacourse/omok/model/entity/Point.kt
@@ -1,0 +1,3 @@
+package woowacourse.omok.model.entity
+
+data class Point(val x: Int, val y: Int)
