+++ app/src/main/java/woowacourse/omok/model/entity/Stone.kt
@@ -1,0 +1,6 @@
+package woowacourse.omok.model.entity
+
+data class Stone(
+    val point: Point,
+    val stoneColor: StoneColor,
+)
