+++ app/src/main/java/woowacourse/omok/model/entity/StoneColor.kt
@@ -1,0 +1,6 @@
+package woowacourse.omok.model.entity
+
+enum class StoneColor {
+    BLACK,
+    WHITE,
+}
