+++ app/src/main/java/woowacourse/omok/model/rule/FiveInRowRule.kt
@@ -1,0 +1,3 @@
+package woowacourse.omok.model.rule
+
+object FiveInRowRule : Rule by StonesInRowRule(5)
