+++ app/src/main/java/woowacourse/omok/model/rule/OverSixInRowRule.kt
@@ -1,0 +1,3 @@
+package woowacourse.omok.model.rule
+
+object OverSixInRowRule : Rule by StonesInRowRule(6)
