+++ app/src/main/java/woowacourse/omok/model/rule/Rule.kt
@@ -1,0 +1,7 @@
+package woowacourse.omok.model.rule
+
+import woowacourse.omok.model.board.Board
+
+interface Rule {
+    fun check(board: Board): Boolean
+}
