+++ app/src/main/java/woowacourse/omok/view/input/InputView.kt
@@ -1,0 +1,5 @@
+package woowacourse.omok.view.input
+
+interface InputView {
+    fun getStonePoint(): Pair<Int, Int>
+}
