+++ app/src/main/java/woowacourse/omok/view/output/OutputView.kt
@@ -1,0 +1,20 @@
+package woowacourse.omok.view.output
+
+import woowacourse.omok.model.board.Board
+import woowacourse.omok.model.entity.StoneColor
+
+interface OutputView {
+    fun printStartGuide()
+
+    fun printTurn(
+        board: Board,
+        stoneColor: StoneColor,
+    )
+
+    fun printWinner(
+        board: Board,
+        stoneColor: StoneColor,
+    )
+
+    fun printInAppropriatePlace(message: String)
+}
