--- src/main/kotlin/omok/controller/OmokController.kt
@@ -1,16 +1,0 @@
-package omok.controller
-
-import omok.model.OmokGame
-import omok.view.input.InputView
-import omok.view.output.OutputView
-
-class OmokController(
-    private val inputView: InputView,
-    private val outputView: OutputView,
-    private val omokGame: OmokGame,
-) {
-    fun startOmok() {
-        outputView.printStartGuide()
-        omokGame.run(inputView::getStonePoint, outputView::printTurn, outputView::printWinner, outputView::printInAppropriatePlace)
-    }
-}
