--- src/main/kotlin/omok/model/board/BoardState.kt
@@ -1,11 +1,0 @@
-package omok.model.board
-
-sealed interface BoardState
-
-class Duplicated(val message: String) : BoardState
-
-class OutOfRange(val message: String) : BoardState
-
-class Full(val message: String) : BoardState
-
-class Success(val board: Board) : BoardState
