--- src/main/kotlin/omok/model/entity/Point.kt
@@ -1,3 +1,0 @@
-package omok.model.entity
-
-data class Point(val x: Int, val y: Int)
