--- src/main/kotlin/omok/model/entity/Stone.kt
@@ -1,6 +1,0 @@
-package omok.model.entity
-
-data class Stone(
-    val point: Point,
-    val stoneColor: StoneColor,
-)
