--- src/main/kotlin/omok/model/entity/StoneColor.kt
@@ -1,6 +1,0 @@
-package omok.model.entity
-
-enum class StoneColor {
-    BLACK,
-    WHITE,
-}
