--- src/main/kotlin/omok/model/rule/FiveInRowRule.kt
@@ -1,3 +1,0 @@
-package omok.model.rule
-
-object FiveInRowRule : Rule by StonesInRowRule(5)
