--- src/main/kotlin/omok/model/rule/OverSixInRowRule.kt
@@ -1,3 +1,0 @@
-package omok.model.rule
-
-object OverSixInRowRule : Rule by StonesInRowRule(6)
