--- src/main/kotlin/omok/model/rule/Rule.kt
@@ -1,7 +1,0 @@
-package omok.model.rule
-
-import omok.model.board.Board
-
-interface Rule {
-    fun check(board: Board): Boolean
-}
