--- src/main/kotlin/omok/view/input/InputView.kt
@@ -1,5 +1,0 @@
-package omok.view.input
-
-interface InputView {
-    fun getStonePoint(): Pair<Int, Int>
-}
