--- src/main/kotlin/omok/view/output/OutputView.kt
@@ -1,20 +1,0 @@
-package omok.view.output
-
-import omok.model.board.Board
-import omok.model.entity.StoneColor
-
-interface OutputView {
-    fun printStartGuide()
-
-    fun printTurn(
-        board: Board,
-        stoneColor: StoneColor,
-    )
-
-    fun printWinner(
-        board: Board,
-        stoneColor: StoneColor,
-    )
-
-    fun printInAppropriatePlace(message: String)
-}
